Subjects, Semester, Lesson with some bugs

+++ src/main/java/com/peter/kist/repository/LessonRepository.java
@@ -3,5 +3,5 @@ package com.peter.kist.repository;
 import com.peter.kist.model.Lesson;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LessonRepository extends JpaRepository<Lesson, Long> {
+public interface LessonRepository extends JpaRepository<Lesson, Integer> {
 }
